DataBase

+++ src/com/startjava/leasson_5/create_db.sql
@@ -8,16 +8,16 @@ CREATE DATABASE Jaegers;
 
 \c jaegers;
 
-create table Jaegers(
-  id BIGINT NOT NULL PRIMARY KEY,
-  modelName VARCHAR(100),
-  mark VARCHAR(100),
-  height FLOAT,
-  weight FLOAT,
-  status VARCHAR(100),
-  origin VARCHAR(100),
-  launch VARCHAR(100),
-  kaijuKill INTEGER
+create table Jaegers (
+    id SERIAL NOT NULL PRIMARY KEY,
+    modelName VARCHAR(100),
+    mark VARCHAR(100),
+    height FLOAT,
+    weight FLOAT,
+    status VARCHAR(100),
+    origin VARCHAR(100),
+    launch VARCHAR(100),
+    kaijuKill INTEGER
 );
 
 \ir init_db.sql
